Refactor DRP calendar header button
